used context api to manage state of movie search
