+++ client/src/utils/actions.js
@@ -1,1 +1,4 @@
 export const UPDATE_CURRENT_CATEGORY = 'UPDATE_CURRENT_CATEGORY';
+export const UPDATE_SEARCH_INPUT = 'UPDATE_SEARCH_INPUT';
+export const UPDATE_MOVIE = 'UPDATE_MOVIE';
+export const UPDATE_SEARCHED_BOOKS = 'UPDATE_SEARCHED_BOOKS';
