+++ client/src/utils/reducers.js
@@ -1,6 +1,10 @@
 import { useReducer } from 'react';
 
-import { UPDATE_CURRENT_CATEGORY } from './actions';
+import { 
+    UPDATE_CURRENT_CATEGORY,
+    UPDATE_SEARCH_INPUT, 
+    UPDATE_MOVIE, 
+    UPDATE_SEARCHED_BOOKS } from './actions';
 
 export const reducer = (state, action) => {
     switch (action.type) {
@@ -8,6 +12,21 @@ export const reducer = (state, action) => {
             return {
                 ...state,
                 currentCategory: action.currentCategory
+            };
+        case UPDATE_SEARCH_INPUT:
+            return {
+                ...state,
+                searchInput: action.searchInput
+            };
+        case UPDATE_MOVIE:
+            return {
+                ...state,
+                movie: action.movie
+            };
+        case UPDATE_SEARCHED_BOOKS:
+            return {
+                ...state,
+                searchedBooks: action.searchedBooks
             };
         default:
             return state;
